# src/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Kubernetes Namespace Provisioner API
  description: |
    TypeScript implementation of the Eclipse Che server's Kubernetes namespace provisioning API.
    
    This API provides endpoints for provisioning and managing Kubernetes namespaces where 
    authenticated users can create workspaces.
    
    ## Original Java Implementation
    This API is based on the Eclipse Che server endpoint:
    - Java Class: `org.eclipse.che.workspace.infrastructure.kubernetes.api.server.KubernetesNamespaceService`
    - Endpoint: `POST /kubernetes/namespace/provision`
    
    ## Authentication
    The API supports two authentication methods:
    1. **Bearer Token**: Format `userid:username` (e.g., `Bearer user123:johndoe`)
    2. **Basic Auth**: Username and userid as password (e.g., `johndoe:user123`)
    
    ## Features
    - ✅ Provision Kubernetes namespaces with configurable name templates
    - ✅ List all managed namespaces
    - ✅ Automatic namespace creation if not exists
    - ✅ Namespace metadata with status and attributes
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Eclipse Public License 2.0
    url: https://www.eclipse.org/legal/epl-2.0/
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /scm/resolve:
    get:
      tags:
        - scm
      summary: Resolve file content from SCM repository
      description: |
        Get file content by specific repository URL and filename.
        
        This endpoint resolves and retrieves file content from various SCM providers:
        - GitHub
        - GitLab
        - Bitbucket
        - Any Git-based repository with public access
        
        The resolver automatically detects the SCM provider based on the repository URL
        and uses the appropriate API or raw content URL to fetch the file.
        
        **Common use cases**:
        - Fetching devfile.yaml from repositories
        - Reading configuration files
        - Accessing repository metadata files
        
        **Note**: For private repositories, ensure proper OAuth tokens are configured.
      operationId: resolveScmFile
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - name: repository
          in: query
          description: Repository URL (GitHub, GitLab, Bitbucket, or any Git repository)
          required: true
          schema:
            type: string
          examples:
            github:
              value: https://github.com/eclipse-che/che-server
              summary: GitHub repository
            gitlab:
              value: https://gitlab.com/user/project
              summary: GitLab repository
            raw:
              value: https://raw.githubusercontent.com/eclipse/che/main/devfile.yaml
              summary: Direct raw URL
        - name: file
          in: query
          description: File path or name within the repository
          required: true
          schema:
            type: string
          examples:
            devfile:
              value: devfile.yaml
              summary: Devfile in root
            nested:
              value: .devfile/devfile.yaml
              summary: Nested devfile
            readme:
              value: README.md
              summary: README file
      responses:
        '200':
          description: File content successfully retrieved
          content:
            text/plain:
              schema:
                type: string
              example: |
                schemaVersion: 2.1.0
                metadata:
                  name: che-server
                components:
                  - name: tools
                    container:
                      image: quay.io/devfile/universal-developer-image:latest
        '400':
          description: Missing required parameters or invalid repository URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingRepository:
                  value:
                    error: Bad Request
                    message: Repository parameter is required
                missingFile:
                  value:
                    error: Bad Request
                    message: File parameter is required
                noResolver:
                  value:
                    error: Bad Request
                    message: Cannot find suitable file resolver for the provided URL
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Requested file not found in repository
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Not Found
                message: Requested file not found in repository
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type or category
          example: Unauthorized
        message:
          type: string
          description: Detailed error message
          example: Authorization header is required
        details:
          type: string
          description: Additional error details (only in development mode)
          example: Missing Authorization header in request
    ExtendedError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            stack:
              type: string
              description: Stack trace (only in development mode)
              example: 'Error: Not able to find namespace\n    at NamespaceProvisioner.provision...'
  responses:
    Unauthorized:
      description: Unauthorized - Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingAuth:
              value:
                error: Unauthorized
                message: Authorization header is required
            invalidAuth:
              value:
                error: Unauthorized
                message: Invalid authorization credentials
    InternalServerError:
      description: Internal server error occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExtendedError'
          examples:
            namespaceNotFound:
              value:
                error: Internal Server Error
                message: Not able to find namespace che-johndoe
            kubernetesError:
              value:
                error: Internal Server Error
                message: Failed to create namespace
                details: Kubernetes API error - Forbidden
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: userid:username
      description: |
        Bearer token authentication using the format `userid:username`.
        
        Example: `Authorization: Bearer user123:johndoe`
        
        In production, this should be replaced with proper JWT tokens.

    BasicAuth:
      type: http
      scheme: basic
      description: |
        Basic authentication using username and userid.
        
        - Username: user's username (e.g., `johndoe`)
        - Password: user's ID (e.g., `user123`)
        
        Example: `Authorization: Basic am9obmRvZTp1c2VyMTIz` (base64 of `johndoe:user123`)
security:
  - BearerAuth: []
  - BasicAuth: []